+++ primitives/src/circuit/signature/schnorr.rs
@@ -7,7 +7,7 @@
 //! Circuit implementation of a Schnorr signature scheme.
 
 use crate::{
-    circuit::rescue::RescueGadget,
+    circuit::rescue::RescueNativeGadget,
     constants::CS_ID_SCHNORR,
     rescue::RescueParameter,
     signatures::schnorr::{Signature, VerKey},
@@ -185,7 +185,8 @@ where
         ];
         chal_input.extend(msg);
 
-        let challenge = self.rescue_sponge_with_padding(&chal_input, 1)?[0];
+        let challenge =
+            RescueNativeGadget::<F>::rescue_sponge_with_padding(self, &chal_input, 1)?[0];
         let c_bits = self.unpack(challenge, field_bit_len::<F>())?;
         Ok(c_bits[..challenge_bit_len::<F>()].to_vec())
     }